Refactor:recycle:: #16 Division of Data File
- 데이터 파일 구분
